requirements.txt
Flask==3.0.0
azure-storage-blob==12.19.0
azure-data-tables==12.5.0
Werkzeug==3.0.1
opencv-python>=4.9.0.80
Pillow>=10.3.0
requests>=2.31.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
